This patch add the notion of groups to the project
A group is just a container of dashboards.
This new notion introduce the possibility of filter the dashboards
by group, linking a dashboard with the others in the group, etc...
Those features will be commited later.
In this commit :
	- Add a group
	- Edit a group
	- Delete a group
There is also the script that create the new sql table and, if the base
is already created, a script that add the column "group_id"
in the dashboard table.