Patch : import and export dashboards in JSON
The user can export a simple dashboard or a selection of dashboards.
The user can also import a JSON file that contains one or several
dashboards which will be added into the group that is currently filtered.
(If there is no filter, the dashboards will be added into the default group)