Patch:reorder graphs & lines (SQL script attached)
This patch introduce to the user the ability of reordering graphs and lines.
To do so, a column was added in the lines table to introduce a weight.
Up and down arrow are now displayed so that the user can reorder by clicking.
The "drag and drop" functionnality is also allowed.
Finally, a loader screen makes the user waiting for the action to complete.
(The drag and drop functionnality needs the latest version of jquery-ui
commited previously)