Javascript refresh instead of meta one
The meta refresh has been replaced by a javascript one so that
it can be disabled. A graph zoom is lost when the page is refreshed,
so the need is to stop the refresh rate during the analysis of the graphs
when they are zoomed.
A button to edit the graph has been added and also a button to view in large
the graph with it current zoom.
The left-margin of a graph isn't the same if the graph has or not
a title for the y-axis, so this margin has been parameterized.