Import test on the name instead of user_id
To know if the file contains a dashboard or an array of dashboard,
we tested the field "user_id".
The fact is that we take the user_id of the user in session, so this field
is no longer compulsory in the imported dashboard. Therefore we have to change
the tested field to know the desired information (if the dashboard is
in an array or not)