[#61435466] Fix SSL offload http redirect issue
From the client to the Load balancer https is in effect but from
the A10 LB to the apache server it's http. The application detects
this and uses http://tattle.kabam.com as the base url during redirects
which of course does not exist. Unfortunately the A10 LB does not seem
to send any indicator of doing SSL offloading via server headers, etc.