Define an offset for the alert control
With different ways to record metrics (batch, etc...),
these metrics obviously aren't set in the same second.
Therefore when we have a threshold alert with a complex target,
graphite computes its result with incomplete data
that may lead to a false positive alert.

To prevent this false positive alert, we have defined an offset
(set by default to one minute) that insure graphite to have the whole data
in order to make its calculation.

This commit resolves issue #55