Adding quick full text search for Dashboards
1. An input is added on the page which allows user to filter Dashboards
by Name/Description, Name/Description/Vtitle/Area of its Graphs and also
Alias/Target of its Lines. i.e if one or more Graphs/Lines pass the
filter while the Dashboard doesn't, the Dashboard will not be filtered.
2. The Dashboards table will be automatically updated by text entered in
the filter
3. The number of Graphs which passed the filter will be shown in the
left bubble and the number of Lines which passed the filter will be
shown in the right bubble in the column "Name" of the Dashboard
4. The column containing a Name/Description that passes the filter will
be highlighted